+++ frontend/src/metabase/components/DateAllOptionsWidget/DateAllOptionsWidget.tsx
@@ -2,7 +2,6 @@ import React, { useState } from "react";
 import { t } from "ttag";
 import cx from "classnames";
 
-import { dateParameterValueToMBQL } from "metabase/parameters/utils/mbql";
 import DatePicker from "metabase/query_builder/components/filters/pickers/DatePicker/DatePicker";
 import { filterToUrlEncoded } from "metabase/parameters/utils/date-formatting";
 
@@ -10,6 +9,8 @@ import {
   WidgetRoot,
   UpdateButton,
 } from "metabase/parameters/components/widgets/Widget.styled";
+
+import { dateParameterValueToMBQL } from "metabase-lib/lib/parameters/utils/mbql";
 
 // Use a placeholder value as field references are not used in dashboard filters
 const noopRef = null;
